Examples.v: progress
Better presentation of cost refining.

The refined expression (before simplification) now only appears in one
place. The subgoal corresponding to the cleanup of the cost function is readable
and doesn't mention explicitly the raw cost function.

The context isn't polluted either, it only contains a definition for the cost
function being elaborated, in the subgoals corresponding to the proof of the
program. The subgoals for "cost function is positive" and "cost function is
dominated by ..." only see the simplified cost function.