Clean up the definition and usage of [cumul]
Defining a custom [cumul] for the product case, and [cumul_with] as a wrapper of
[cumul] indeeds gives cute specifications in _some_ cases. However it does not
really scale to more complicated cases (e.g. when the function in the sum
depends on the upper bound). Moreover, it duplicates lemmas, and introduces
indirections which make it harder to understand the specs.

Sneaked in: the order of arguments of [cumul] changed, from [cumul f lo hi] to
[cumul lo hi f]. Hopefully this is more close to the math notation, where one
reads the index of the sum before reading its body.