Change the statement of dominated lemmas involving products
This changes the lemmas so that they can be used by "apply". The
essence of why this is needed is because Coq is able to solve a
unification problem such as "?f a b = a + b", but not "?f (a, b) = a +
b". So having a non-curried function in the conclusion of the lemma,
to be unified with the goal usually does not work.